simplify (and speed up) no-encoding case
We always return a response, just not always an encoded one. All we
really check for is, does the client accept one of the encodings we
make, or does it not. So it is pointless to scan for the `identity`
encoding: it is simply among the encodings we do not make.